requirements.txt
astropy==2.0.2
attrs==17.3.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.1.4
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
html5lib==1.0.1
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.33.12
jupyterlab-launcher==0.11.2
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.6.0
numpy==1.13.3
pandas==0.21.0
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pluggy==0.6.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.5.2
pyaml==17.12.1
Pygments==2.2.0
pyparsing==2.2.0
pytest==3.3.1
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
scandir==1.9.0
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
subprocess32==3.5.2
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
